fix: update test import paths after module reorganization
